0.3.3 Release (#47)
* #38 Support comments for Scenario Outlines
Added support for both // and # comment styles. // being native to javascript is probably the easiest way to add comments. # is because this is the characer used in Gherkin

* #38 Support comments for Scenario Outlines
Added support for both // and # comment styles. // being native to javascript is probably the easiest way to add comments. # is because this is the characer used in Gherkin

* updating release notes

* #46 Throw exception when using Async for Feature/Scenario/ScenarioOutline/Background
#45 Background doesn't support async on multiple scenarios

* Fixing issue with backgrounds
Backgrounds would fail if and error occurs during execution